+++ lib/braintree/credit_card_gateway.rb
@@ -81,9 +81,9 @@ module Braintree
       billing_address_params = AddressGateway._shared_signature
       options = [:make_default, :verification_merchant_account_id, :verify_card, :verification_amount, :venmo_sdk_session, :fail_on_duplicate_payment_method, :verification_account_type]
       signature = [
-        :billing_address_id, :cardholder_name, :cvv, :device_session_id, :expiration_date,
-        :expiration_month, :expiration_year, :number, :token, :venmo_sdk_payment_method_code,
-        :device_data, :fraud_merchant_id, :payment_method_nonce,
+        :billing_address_id, :cardholder_name, :cvv, :expiration_date, :expiration_month,
+        :expiration_year, :number, :token, :venmo_sdk_payment_method_code, :device_data,
+        :payment_method_nonce,
         {:external_vault => [:network_transaction_id]},
         {:options => options},
         {:billing_address => billing_address_params}