+++ lib/braintree/payment_method_gateway.rb
@@ -162,9 +162,9 @@ module Braintree
         ],
       ]
       signature = [
-        :billing_address_id, :cardholder_name, :cvv, :device_session_id, :expiration_date,
-        :expiration_month, :expiration_year, :number, :token, :venmo_sdk_payment_method_code,
-        :device_data, :fraud_merchant_id, :payment_method_nonce,
+        :billing_address_id, :cardholder_name, :cvv, :expiration_date, :expiration_month,
+        :expiration_year, :number, :token, :venmo_sdk_payment_method_code, :device_data,
+        :payment_method_nonce,
         {:options => options},
         {:billing_address => billing_address_params}
       ]